both versions of the Lloyd algorithm (Matlab and Python)
The matlab version has one tweak from Matthew's version - the distortion function had a abs around the different when comparing the m+1 to the mth distortion, so I changed that to a square (based on what the algorithm outlines).  Results with the square are consistent with what we should have.